Add prefixes to json to denote type to deserialize to
