+++ src/main/java/animatedledstrip/animationutils/AnimationDataUtils.kt
@@ -29,7 +29,6 @@ import animatedledstrip.colors.ColorContainerSerializer
 import animatedledstrip.colors.ccpresets.CCBlack
 import animatedledstrip.utils.parseHex
 import com.google.gson.GsonBuilder
-import java.nio.charset.Charset
 
 /* JSON Parser */
 val gson = GsonBuilder()
@@ -351,10 +350,3 @@ fun Animation.isNonRepetitive() =
 
 fun AnimationData.isContinuous(): Boolean = continuous
     ?: !animation.isNonRepetitive()
-
-fun AnimationData.json(): ByteArray = gson.toJson(this).toByteArray(Charset.forName("utf-8"))
-
-fun ByteArray?.jsonToAnimationData(size: Int): AnimationData {
-    checkNotNull(this)
-    return gson.fromJson(this.toString(Charset.forName("utf-8")).take(size), AnimationData::class.java)
-}
