+++ src/test/java/UtilsTest.kt
@@ -23,9 +23,11 @@ package animatedledstrip.test
  */
 
 
+import animatedledstrip.animationutils.*
 import animatedledstrip.colors.ColorContainer
 import animatedledstrip.colors.ccpresets.CCBlack
 import animatedledstrip.colors.ccpresets.CCBlue
+import animatedledstrip.leds.StripInfo
 import animatedledstrip.utils.*
 import org.junit.Test
 import kotlin.test.assertFailsWith
@@ -70,5 +72,58 @@ class UtilsTest {
     fun testToColorContainer() {
         assertTrue { 0xFF.toColorContainer() == ColorContainer(0xFF) }
         assertTrue { 0xFFFFFFL.toColorContainer() == ColorContainer(0xFFFFFF) }
+    }
+
+    @Test
+    fun testAnimationDataJson() {
+        val testAnimation = AnimationData().animation(Animation.STACK)
+            .color(ColorContainer(0xFF, 0xFFFF).prepare(5), index = 0)
+            .color(0xFF, index = 1)
+            .color(0xFF, index = 2)
+            .color(0xFF, index = 3)
+            .color(0xFF, index = 4)
+            .continuous(true)
+            .delay(50)
+            .direction(Direction.FORWARD)
+            .id("TEST")
+            .spacing(5)
+
+        val testBytes = testAnimation.json()
+
+        val testAnimation2 = testBytes.jsonToAnimationData(testBytes.size)
+
+        assertTrue { testAnimation == testAnimation2 }
+
+        assertFailsWith<IllegalStateException> {
+            val nullBytes: ByteArray? = null
+            nullBytes.jsonToAnimationData(0)
+        }
+    }
+
+    @Test
+    fun testStripInfoJson() {
+        val info1 = StripInfo()
+        val infoBytes = info1.json()
+
+        val info2 = infoBytes.jsonToStripInfo(infoBytes.size)
+
+        assertTrue { info1 == info2 }
+
+        assertFailsWith<IllegalStateException> {
+            val nullBytes: ByteArray? = null
+            nullBytes.jsonToStripInfo(0)
+        }
+    }
+
+    @Test
+    fun testGetDataTypePrefix() {
+        val info1 = StripInfo()
+        val infoBytes = info1.json()
+        assertTrue { infoBytes.getDataTypePrefix() == "INFO" }
+
+        val animTest = AnimationData()
+        val animBytes = animTest.json()
+        assertTrue { animBytes.getDataTypePrefix() == "DATA" }
+
     }
 }
